pyproject.toml
[tool.poetry]
name = "workflow-nuvs"
version = "0.0.0"
description = "A workflow for identifying novel viruses in Virtool."
authors = ["Ian Boyes", "Blake Smith"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
biopython = "^1.79"
virtool-workflow = "^5.0.0"
rust = "^0.1.1"
nuvs-rust = {path = "target/wheels/nuvs_rust-0.1.0-cp310-cp310-manylinux_2_31_x86_64.whl"}

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
pytest-aiohttp = "^0.3.0"
pytest-xdist = "^2.3.0"
pytest-regressions = "^2.2.0"
pydantic-factories = "^1.17.0"
maturin = "^0.13.1"
black = "^23.1.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0","maturin>=0.14,<0.15"]
build-backend = "maturin"
